style: fix clippy warnings in memory and sandbox crates
Resolve the -D warnings failures from the nightly CI clippy job across
crates/memory (needless_range_loop, let_and_return, useless_conversion,
unnecessary_sort_by, map_err closures, manual_flatten, char-pattern
matching) and crates/sandbox (io_other_error, manual_repeat_n,
unnecessary_cast, vec_init_then_push, manual_checked_ops, manual_strip,
manual_pattern_char_comparison, manual_inspect, redundant_closure,
useless_borrows_in_formatting, dead code). Platform API helpers
(CgroupV2Controller, apply_cgroup_for_self) keep #[allow(dead_code)]
rather than deletion.

+++ crates/sandbox/src/seccomp/syscalls.rs
@@ -710,7 +710,7 @@ fn linux_lookup(name: &str) -> Option<i64> {
         "fchmodat2" => libc::SYS_fchmodat2,
         "mseal" => libc::SYS_mseal,
         _ => return fallback_x86_64_lookup(name),
-    } as i64)
+    })
 }
 
 /// Fallback lookup for syscalls that may not be defined in an older `libc`
